requirements.txt
openai
faiss-cpu
tqdm
PyPDF2
python-dotenv
streamlit
langchain
sentence-transformers